Add simple Vanity address support
You can now create custom vanity addresses in just a couple clicks, choose your prefix and mine away!

Currently the generator is quite slow, so it's only effective at generating up to 2 letters of prefix. E.G: "SCC" as an address, the "S" being excluded from the calculation since every address begins with S.

Optimizations and multi-threaded generation is planned for the future to increase the viability of vanity addresses.